+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/export-default-interface/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+export default interface T {
+  method1(): void;
+}