+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/type-assertion-in-function/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+function assertsString(x: any): asserts x {
+  return;
+}