+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/type-assertion-with-guard-in-function/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+function assertsStringGuard(x: any): asserts x is string {
+  return;
+}