+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/type-guard-in-arrow-function/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+const isString = (x: any): x is string => {
+  return typeof x === 'string';
+};