+++ packages/ast-spec/src/legacy-fixtures/expressions/fixtures/optional-call-expression-type-arguments/fixture.ts
@@ -1,0 +1,4 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+foo?.bar<A>();
+foo?.bar<number>();