+++ packages/eslint-plugin/src/rules/no-import-type-side-effects.ts
@@ -28,6 +28,10 @@ export default util.createRule<Options, MessageIds>({
       'ImportDeclaration[importKind!="type"]'(
         node: TSESTree.ImportDeclaration,
       ): void {
+        if (node.specifiers.length === 0) {
+          return;
+        }
+
         const specifiers: TSESTree.ImportSpecifier[] = [];
         for (const specifier of node.specifiers) {
           if (