+++ packages/eslint-plugin/src/rules/no-unsafe-assignment.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
-import * as tools from 'ts-api-tools';
+import * as tools from 'ts-api-utils';
 import type * as ts from 'typescript';
 
 import * as util from '../util';