+++ packages/eslint-plugin/src/rules/strict-boolean-expressions.ts
@@ -3,7 +3,7 @@ import type {
   TSESTree,
 } from '@typescript-eslint/utils';
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
-import * as tools from 'ts-api-tools';
+import * as tools from 'ts-api-utils';
 import * as ts from 'typescript';
 
 import * as util from '../util';
@@ -16,6 +16,7 @@ export type Options = [
     allowNullableBoolean?: boolean;
     allowNullableString?: boolean;
     allowNullableNumber?: boolean;
+    allowNullableEnum?: boolean;
     allowAny?: boolean;
     allowRuleToRunWithoutStrictNullChecksIKnowWhatIAmDoing?: boolean;
   },
@@ -32,6 +33,7 @@ export type MessageId =
   | 'conditionErrorNullableNumber'
   | 'conditionErrorObject'
   | 'conditionErrorNullableObject'
+  | 'conditionErrorNullableEnum'
   | 'noStrictNullCheck'
   | 'conditionFixDefaultFalse'
   | 'conditionFixDefaultEmptyString'
@@ -65,6 +67,7 @@ export default util.createRule<Options, MessageId>({
           allowNullableBoolean: { type: 'boolean' },
           allowNullableString: { type: 'boolean' },
           allowNullableNumber: { type: 'boolean' },
+          allowNullableEnum: { type: 'boolean' },
           allowAny: { type: 'boolean' },
           allowRuleToRunWithoutStrictNullChecksIKnowWhatIAmDoing: {
             type: 'boolean',
@@ -104,6 +107,9 @@ export default util.createRule<Options, MessageId>({
       conditionErrorNullableObject:
         'Unexpected nullable object value in conditional. ' +
         'An explicit null check is required.',
+      conditionErrorNullableEnum:
+        'Unexpected nullable enum value in conditional. ' +
+        'Please handle the nullish/zero/NaN cases explicitly.',
       noStrictNullCheck:
         'This rule requires the `strictNullChecks` compiler option to be turned on to function correctly.',
 
@@ -139,6 +145,7 @@ export default util.createRule<Options, MessageId>({
       allowNullableBoolean: false,
       allowNullableString: false,
       allowNullableNumber: false,
+      allowNullableEnum: true,
       allowAny: false,
       allowRuleToRunWithoutStrictNullChecksIKnowWhatIAmDoing: false,
     },
@@ -719,6 +726,35 @@ export default util.createRule<Options, MessageId>({
         return;
       }
 
+      // nullable enum
+      if (is('nullish', 'number', 'enum') || is('nullish', 'string', 'enum')) {
+        if (!options.allowNullableEnum) {
+          if (isLogicalNegationExpression(node.parent!)) {
+            context.report({
+              node,
+              messageId: 'conditionErrorNullableEnum',
+              fix: util.getWrappingFixer({
+                sourceCode,
+                node: node.parent,
+                innerNode: node,
+                wrap: code => `${code} == null`,
+              }),
+            });
+          } else {
+            context.report({
+              node,
+              messageId: 'conditionErrorNullableEnum',
+              fix: util.getWrappingFixer({
+                sourceCode,
+                node,
+                wrap: code => `${code} != null`,
+              }),
+            });
+          }
+        }
+        return;
+      }
+
       // any
       if (is('any')) {
         if (!options.allowAny) {
@@ -754,6 +790,7 @@ export default util.createRule<Options, MessageId>({
       | 'number'
       | 'truthy number'
       | 'object'
+      | 'enum'
       | 'any'
       | 'never';
 
@@ -780,7 +817,7 @@ export default util.createRule<Options, MessageId>({
       // If incoming type is either "true" or "false", there will be one type
       // object with intrinsicName set accordingly
       // If incoming type is boolean, there will be two type objects with
-      // intrinsicName set "true" and "false" each because of ts-api-tools.unionTypeParts()
+      // intrinsicName set "true" and "false" each because of ts-api-utils.unionTypeParts()
       if (booleans.length === 1) {
         tools.isBooleanLiteralType(booleans[0], true)
           ? variantTypes.add('truthy boolean')
@@ -813,6 +850,12 @@ export default util.createRule<Options, MessageId>({
         } else {
           variantTypes.add('number');
         }
+      }
+
+      if (
+        types.some(type => tools.isTypeFlagSet(type, ts.TypeFlags.EnumLike))
+      ) {
+        variantTypes.add('enum');
       }
 
       if (