+++ packages/shared-fixtures/fixtures/javascript/function/return-multiline-sequence.src.js
@@ -1,7 +1,0 @@
-function foo (x, y, z) {
-  return (
-    x,
-    y,
-    z
-  );
-}