+++ packages/shared-fixtures/fixtures/javascript/generators/async-generator-function.src.js
@@ -1,4 +1,0 @@
-
-async function *foo() {
-
-}