+++ packages/shared-fixtures/fixtures/typescript/basics/abstract-class-with-abstract-properties.src.ts
@@ -1,4 +1,0 @@
-abstract class Foo {
-    abstract bar;
-    abstract baz: number;
-}