+++ packages/shared-fixtures/fixtures/typescript/basics/class-with-two-methods-computed-constructor.src.ts
@@ -1,7 +1,0 @@
-class A {
-  "constructor"<T>(): number {
-  }
-
-  ["constructor"]<T>(): number {
-  }
-};