+++ packages/shared-fixtures/fixtures/typescript/basics/export-named-enum-computed-number.src.ts
@@ -1,3 +1,0 @@
-export enum Foo {
-    [1],
-}