+++ packages/shared-fixtures/fixtures/typescript/basics/function-anynomus-with-return-type.src.ts
@@ -1,2 +1,0 @@
-var obj = function (): void {
-};