+++ packages/shared-fixtures/fixtures/typescript/declare/enum.src.ts
@@ -1,4 +1,0 @@
-declare enum Foo {
-    Bar,
-    Baz
-}