+++ packages/shared-fixtures/fixtures/typescript/types/intersection-type.src.ts
@@ -1,1 +1,0 @@
-type LinkedList<T> = T & { next: LinkedList<T> };