+++ packages/typescript-estree/tests/fixtures/projectTrue/nested/included.ts
@@ -1,0 +1,1 @@
+const b = true;