+++ packages/website/src/theme/MDXComponents/RuleAttributes.tsx
@@ -1,3 +1,4 @@
+import Link from '@docusaurus/Link';
 import type { RuleMetaDataDocs } from '@site/../utils/dist/ts-eslint/Rule';
 import { useRulesMeta } from '@site/src/hooks/useRulesMeta';
 import React from 'react';
@@ -35,18 +36,15 @@ export function RuleAttributes({ name }: { name: string }): React.ReactNode {
       children: (
         <>
           Extending{' '}
-          <a href={`/docs/linting/configs#${recommendation}`} target="_blank">
+          <Link to={`/linting/configs#${recommendation}`} target="_blank">
             <code className={styles.code}>
               "plugin:@typescript-eslint/{recommendation}"
             </code>
-          </a>{' '}
+          </Link>{' '}
           in an{' '}
-          <a
-            href="https://eslint.org/docs/latest/user-guide/configuring/configuration-files#extending-configuration-files"
-            target="_blank"
-          >
+          <Link href="https://eslint.org/docs/latest/user-guide/configuring/configuration-files#extending-configuration-files">
             ESLint configuration
-          </a>{' '}
+          </Link>{' '}
           enables this rule.
         </>
       ),
@@ -59,12 +57,9 @@ export function RuleAttributes({ name }: { name: string }): React.ReactNode {
       children: (
         <>
           Some problems reported by this rule are automatically fixable by the{' '}
-          <a
-            href="https://eslint.org/docs/latest/user-guide/command-line-interface#--fix"
-            target="_blank"
-          >
+          <Link href="https://eslint.org/docs/latest/user-guide/command-line-interface#--fix">
             <code>--fix</code> ESLint command line option
-          </a>
+          </Link>
           .
         </>
       ),
@@ -77,12 +72,9 @@ export function RuleAttributes({ name }: { name: string }): React.ReactNode {
       children: (
         <>
           Some problems reported by this rule are manually fixable by editor{' '}
-          <a
-            href="https://eslint.org/docs/latest/developer-guide/working-with-rules#providing-suggestions"
-            target="_blank"
-          >
+          <Link href="https://eslint.org/docs/latest/developer-guide/working-with-rules#providing-suggestions">
             suggestions
-          </a>
+          </Link>
           .
         </>
       ),
@@ -95,9 +87,9 @@ export function RuleAttributes({ name }: { name: string }): React.ReactNode {
       children: (
         <>
           This rule requires{' '}
-          <a href="/linting/typed-linting" target="_blank">
+          <Link href="/linting/typed-linting" target="_blank">
             type information
-          </a>{' '}
+          </Link>{' '}
           to run.
         </>
       ),